Cleaned up omniglot, cancer and diabetes

+++ 03_omniglot/analyze_all.py
@@ -21,33 +21,44 @@ import plot_omniglot
 def analyze_all(args: argparse.Namespace):
     # create directory
     exp_dir_path = args.exp_dir_path
-    results_dir_path = os.path.join(exp_dir_path, 'results')
+    results_dir_path = os.path.join(exp_dir_path, 'results_data=100')
     plot_dir = os.path.join(results_dir_path, 'plots')
     os.makedirs(plot_dir, exist_ok=True)
 
-    inf_algs_results_df, inf_algs_num_features_by_num_obs = load_all_inf_alg_results(
-        results_dir_path=results_dir_path)
+    inf_algs_results_df_path = os.path.join(exp_dir_path,
+                                            'inf_algs_results_df.csv')
 
-    inf_algs_results_df.sort_values(
-        by=['log_posterior_predictive'],
-        inplace=True,
-        ascending=False)
+    if not os.path.isfile(inf_algs_results_df_path):
 
-    inf_algs_results_df.to_csv(
-        os.path.join(exp_dir_path, 'inf_algorithms_results_df.csv'),
-        index=False)
+        inf_algs_results_df = load_all_inf_alg_results(
+            results_dir_path=results_dir_path)
+
+        inf_algs_results_df.sort_values(
+            by=['log_posterior_predictive'],
+            inplace=True,
+            ascending=False)
+
+        inf_algs_results_df.to_csv(
+            inf_algs_results_df_path,
+            index=False)
+
+    else:
+        inf_algs_results_df = pd.read_csv(inf_algs_results_df_path,
+                                          index_col=None)
+
+    print(f"Runs per algorithm:\n"
+          f"{inf_algs_results_df.groupby('inference_alg').size()}")
 
     plot_omniglot.plot_analyze_all_algorithms_results(
         inf_algorithms_results_df=inf_algs_results_df,
-        inf_algs_num_features_by_num_obs=inf_algs_num_features_by_num_obs,
         plot_dir=plot_dir)
 
 
 def load_all_inf_alg_results(results_dir_path: str,
-                             ) -> Tuple[pd.DataFrame, List[np.ndarray]]:
+                             ) -> pd.DataFrame:
 
     inf_algorithms_results_rows = []
-    inf_algorithms_num_features_by_num_obs = []
+    # inf_algorithms_num_features_by_num_obs = []
     run_dirs = [subdir for subdir in os.listdir(results_dir_path)]
 
     # Iterate through each sampling scheme directory
@@ -73,7 +84,7 @@ def load_all_inf_alg_results(results_dir_path: str,
 
         num_features_by_num_obs = stored_data['inference_alg_results'][
                                       'num_dishes_poisson_rate_posteriors'][:, 0]  # remove extra dimension
-        inf_algorithms_num_features_by_num_obs.append(num_features_by_num_obs)
+        # inf_algorithms_num_features_by_num_obs.append(num_features_by_num_obs)
 
         del stored_data
 
@@ -86,7 +97,7 @@ def load_all_inf_alg_results(results_dir_path: str,
     inf_algorithms_results_df['negative_log_posterior_predictive'] = \
         -inf_algorithms_results_df['log_posterior_predictive']
 
-    return inf_algorithms_results_df, inf_algorithms_num_features_by_num_obs
+    return inf_algorithms_results_df  # , inf_algorithms_num_features_by_num_obs
 
 
 if __name__ == '__main__':
